icons error fix
