--- packages/ui/icons.ts
@@ -1,244 +1,0 @@
-import {
-  SunMedium,
-  Moon,
-  Star,
-  Twitter,
-  X,
-  Loader2,
-  ChevronLeft,
-  ChevronRight,
-  ChevronsLeft,
-  ChevronsRight,
-  ChevronUp,
-  ChevronDown,
-  ChevronsUpDown,
-  ArrowUp,
-  ArrowDown,
-  Menu,
-  MoreVertical,
-  MoreHorizontal,
-  Sliders,
-  SlidersHorizontal,
-  Circle,
-  Check,
-  Plus,
-  PlusCircle,
-  Minus,
-  Eye,
-  EyeOff,
-  Trash,
-  Edit,
-  Crop,
-  RefreshCw,
-  Send,
-  Copy,
-  Download,
-  AlertTriangle,
-  Search,
-  Filter,
-  CalendarDays,
-  AlarmClock,
-  User,
-  FileTerminal,
-  Settings,
-  LogOut,
-  Volume2,
-  VolumeX,
-  MessageSquare,
-  CreditCard,
-  Wallet,
-  DollarSign,
-  ShoppingCart,
-  Package,
-  ShoppingBag,
-  BarChart3,
-  UploadCloud,
-  Shirt,
-  Footprints,
-  HardHat,
-  type LucideIcon,
-} from "lucide-react";
-export {
-  type LucideIcon,
-  SunMedium,
-  type LucideProps,
-  Star,
-  ChevronsLeft,
-  ChevronsRight,
-  ChevronsUpDown,
-  ArrowDown,
-  Menu,
-  MoreVertical,
-  Sliders,
-  SlidersHorizontal,
-  PlusCircle,
-  Minus,
-  Eye,
-  EyeOff,
-  Edit,
-  Crop,
-  RefreshCw,
-  Send,
-  Download,
-  AlertTriangle,
-  Search,
-  Filter,
-  CalendarDays,
-  AlarmClock,
-  FileTerminal,
-  LogOut,
-  Volume2,
-  VolumeX,
-  MessageSquare,
-  CreditCard,
-  Wallet,
-  DollarSign,
-  ShoppingCart,
-  Package,
-  ShoppingBag,
-  BarChart3,
-  UploadCloud,
-  Shirt,
-  Footprints,
-  HardHat,
-} from "lucide-react";
-
-export {
-  AlertCircle,
-  ArrowDownNarrowWide,
-  ArrowUp,
-  ArrowUpNarrowWide,
-  BadgeInfo,
-  BellRing,
-  Binary,
-  Bookmark,
-  Brush,
-  Calendar,
-  Check,
-  CheckCircle,
-  CheckCircle2,
-  ChevronDown,
-  ChevronLeft,
-  ChevronRight,
-  ChevronRightSquare,
-  ChevronUp,
-  Circle,
-  Clipboard,
-  Command,
-  Compass,
-  Copy,
-  Diamond,
-  ExternalLink,
-  Flag,
-  GitBranch,
-  Github,
-  Loader,
-  Loader2,
-  Lock,
-  LogIn,
-  Mail,
-  MailCheck,
-  Maximize2,
-  MessageCircle,
-  MessagesSquare,
-  Minimize2,
-  Moon,
-  MoreHorizontal,
-  Pencil,
-  PieChart,
-  Pin,
-  Play,
-  Plus,
-  Reply,
-  RotateCcw,
-  Settings,
-  Settings2,
-  Shapes,
-  Share,
-  Shield,
-  Sparkle,
-  Sparkles,
-  SquareSlash,
-  Sun,
-  Sword,
-  Swords,
-  ThumbsUp,
-  Trash2,
-  TrendingUpIcon,
-  Triangle,
-  Twitter,
-  Unlock,
-  User,
-  X,
-  XCircle,
-  Trash,
-  Text,
-  FlaskConical,
-  History,
-  ArrowLeft,
-  RectangleVertical,
-  SearchIcon,
-} from "lucide-react";
-
-export type Icon = LucideIcon;
-
-export const Icons = {
-  sun: SunMedium,
-  moon: Moon,
-  star: Star,
-  twitter: Twitter,
-  close: X,
-  spinner: Loader2,
-  chevronLeft: ChevronLeft,
-  chevronRight: ChevronRight,
-  chevronsLeft: ChevronsLeft,
-  chevronsRight: ChevronsRight,
-  chevronUp: ChevronUp,
-  chevronDown: ChevronDown,
-  chevronUpDown: ChevronsUpDown,
-  arrowUp: ArrowUp,
-  arrowDown: ArrowDown,
-  menu: Menu,
-  verticalThreeDots: MoreVertical,
-  horizontalThreeDots: MoreHorizontal,
-  verticalSliders: Sliders,
-  horizontalSliders: SlidersHorizontal,
-  circle: Circle,
-  check: Check,
-  add: Plus,
-  addCircle: PlusCircle,
-  remove: Minus,
-  view: Eye,
-  hide: EyeOff,
-  trash: Trash,
-  edit: Edit,
-  crop: Crop,
-  reset: RefreshCw,
-  send: Send,
-  copy: Copy,
-  downlaod: Download,
-  warning: AlertTriangle,
-  search: Search,
-  filter: Filter,
-  alarm: AlarmClock,
-  calendar: CalendarDays,
-  user: User,
-  terminal: FileTerminal,
-  settings: Settings,
-  logout: LogOut,
-  volumne: Volume2,
-  volumneMute: VolumeX,
-  message: MessageSquare,
-  billing: CreditCard,
-  wallet: Wallet,
-  dollarSign: DollarSign,
-  cart: ShoppingCart,
-  product: Package,
-  store: ShoppingBag,
-  chart: BarChart3,
-  upload: UploadCloud,
-  placeholder: Image,
-  clothing: Shirt,
-  shoes: Footprints,
-  accessories: HardHat,
-};
